pyproject.toml
[project]
name = "nwfsc-nmea"
version = "0.1.0"
description = "Extension for pynmea2 to support NOAA Fisheries (NWFSC) proprietary NMEA sentences."
authors = [{name = "Jim-F", email = "jim.f@noaa.gov"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pynmea2>=1.19.0",
    "pyserial>=3.5",
]

[dependency-groups]
dev = [
    "duckdb>=1.5.0",
    "numpy>=2.4.3",
    "pyarrow>=23.0.1",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/nwfsc_nmea"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
